refactor: di directory name change
