+++ app/src/main/java/com/example/goclass/di/RepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.example.goclass.dimodules
+package com.example.goclass.di
 
 import com.example.goclass.repository.Repository
 import org.koin.dsl.module
